# lib/api-zod/src/generated/types/tender.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tender {
  id: number;
  noticeId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  awardedValue?: number | null;
  /** @nullable */
  currency?: string | null;
  procurementStage: string;
  /** @nullable */
  publishedDate?: string | null;
  noticeUrl: string;
  /** none | downloading | downloaded | processed | failed */
  pdfStatus: string;
  unsuccessfulSuppliers: string[];
  createdAt: string;
  updatedAt: string;
}