# lib/api-zod/src/generated/types/tenderDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Tender } from "./tender";

export type TenderDetail = Tender & {
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  rawPdfText?: string | null;
};