# lib/api-zod/src/generated/types/tenderListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Tender } from "./tender";

export interface TenderListResponse {
  tenders: Tender[];
  total: number;
  page: number;
  limit: number;
}