# lib/api-zod/src/generated/types/tenderSearchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TenderSearchRequest {
  /** Search keyword (e.g. "Framework") */
  keyword: string;
  /**
   * Minimum contract value in GBP
   * @nullable
   */
  minValue?: number | null;
  /** Procurement stages to filter by */
  stages?: string[];
}