# lib/api-zod/src/generated/types/tenderStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TenderStats {
  total: number;
  withUnsuccessfulSuppliers: number;
  totalPdfProcessed: number;
  totalPdfFailed: number;
  /** @nullable */
  averageValue?: number | null;
  /** @nullable */
  highestValue?: number | null;
}